+++ medicare1/src/app/emr/models/Medication.ts
@@ -1,0 +1,7 @@
+export interface Medication {
+  name: string;
+  dosage: string;
+  frequency: string;
+  startDate: string;
+  endDate?: string;
+}
