+++ medicare1/src/app/emr/models/Patient.ts
@@ -1,0 +1,15 @@
+import { Report } from './Report';
+import { PatientRecordWithUser } from './PatientRecordWithUser';
+
+export interface Patient {
+  id: string;
+  patientName: string;
+  condition: string;
+  doctor: string;
+  visitDate: string;
+  visitTime: string;
+  status: string;
+  prescription: string;
+  reports: Report[];
+  previousRecords: PatientRecordWithUser[];
+}
