+++ medicare1/src/app/emr/models/PatientRecordWithUser.ts
@@ -1,0 +1,12 @@
+import { User } from './User';
+
+export interface PatientRecordWithUser {
+  id?: string;
+  visitDate: string;
+  visitTime: string;
+  condition: string;
+  doctor: string;
+  prescription: string;
+  status: string;
+  user?: User;
+}
