+++ medicare1/src/app/emr/models/Report.ts
@@ -1,0 +1,8 @@
+export interface Report {
+  name: string;
+  type: string;
+  date: string;
+  file_id?: string;
+  fileType?: string;
+  description?: string;
+}
