+++ medicare1/src/app/emr/models/index.ts
@@ -1,0 +1,8 @@
+// Central export file for all EMR models
+export type { User, GetUserDTO, SigninUserResponse } from './User';
+export type { PatientRecord } from './PatientRecord';
+export type { PatientReportResponse } from './PatientReportResponse';
+export type { Report } from './Report';
+export type { Patient } from './Patient';
+export type { PatientRecordWithUser } from './PatientRecordWithUser';
+export type { Medication } from './Medication';
