+++ medicare1/src/app/services/patientService/patient.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Patient } from '../../emr/emr.component';
 import { environment } from '../../../environments/environment';
+import type { Patient } from '../../emr/models';
 
 @Injectable({
   providedIn: 'root'
